+++ core/src/main/java/com/graphhopper/routing/ev/BikeRoadAccess.java
@@ -27,15 +27,6 @@ public enum BikeRoadAccess {
 
     public static final String KEY = "bike_road_access";
 
-    /**
-     * The access restriction list returned from OSMRoadAccessParser.toOSMRestrictions(TransportationMode.Bike)
-     * does not contain "vehicle" to still allow walking, via 'dismount' (#2981). But to allow
-     * walking via dismount in case of vehicle=private we need bike_road_access == PRIVATE. This
-     * also allows us to limit speed to 5km/h if foot_road_access == YES. See
-     * <a href="https://www.openstreetmap.org/way/1058548816">this way</a>.
-     */
-    public static final List<String> RESTRICTIONS = Arrays.asList("bicycle", "vehicle", "access");
-
     public static EnumEncodedValue<BikeRoadAccess> create() {
         return new EnumEncodedValue<>(BikeRoadAccess.KEY, BikeRoadAccess.class);
     }